Simplify CLI entry point to use new package
